+++ crates/wdl-format/src/config.rs
@@ -32,6 +32,11 @@ fn reorder_sections_default() -> bool {
     false
 }
 
+/// Default for whether to upgrade deprecations.
+fn upgrade_deprecations_default() -> bool {
+    false
+}
+
 /// Configuration for formatting.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Toml, JsonSchema)]
 #[toml(Toml, deny_unknown_fields)]
@@ -61,6 +66,14 @@ pub struct Config {
     #[toml(default = reorder_sections_default())]
     #[schemars(default = "reorder_sections_default")]
     pub reorder_sections: bool,
+    /// Whether to eagerly upgrade deprecated WDL constructs.
+    ///
+    /// Currently this includes changing curly brace command sections (`{}`)
+    /// into heredoc command sections (`<<<>>>`) and changing dollar-style
+    /// placeholders (`${}`) into tilde-style placeholders (`~{}`).
+    #[toml(default = upgrade_deprecations_default())]
+    #[schemars(default = "upgrade_deprecations_default")]
+    pub upgrade_deprecations: bool,
     /// The newline style.
     #[toml(default, FromToml with = parse_string, ToToml with = display)]
     #[schemars(default)]
@@ -76,6 +89,7 @@ impl Default for Config {
             sort_inputs: sort_inputs_default(),
             trailing_commas: trailing_commas_default(),
             reorder_sections: reorder_sections_default(),
+            upgrade_deprecations: upgrade_deprecations_default(),
             newline_style: NewlineStyle::default(),
         }
     }
@@ -121,6 +135,12 @@ impl Config {
     /// Set whether section reordering is enabled.
     pub fn reorder_sections(mut self, reorder_sections: bool) -> Self {
         self.reorder_sections = reorder_sections;
+        self
+    }
+
+    /// Set whether to upgrade deprecations.
+    pub fn upgrade_deprecations(mut self, upgrade_deprecations: bool) -> Self {
+        self.upgrade_deprecations = upgrade_deprecations;
         self
     }
 }