+++ src/commands/inputs.rs
@@ -40,7 +40,7 @@ pub struct Args {
     pub target: Option<String>,
 
     /// Show inputs with non-literal default values.
-    #[arg(long)]
+    #[arg(long, conflicts_with = "type_signatures")]
     pub show_non_literals: bool,
 
     /// Hide inputs with default values.
@@ -50,6 +50,10 @@ pub struct Args {
     /// Generate inputs for all tasks called in the workflow.  
     #[arg(long)]
     pub nested_inputs: bool,
+
+    /// Render all input values as their type signature.
+    #[arg(long, conflicts_with = "show_non_literals")]
+    pub type_signatures: bool,
 
     /// Output the template as a YAML file.
     #[arg(long)]
@@ -100,16 +104,25 @@ pub struct InputProcessor {
 
     /// Whether or not to include defaults.
     hide_defaults: bool,
+
+    /// Whether or not to render all values as type signatures.
+    type_signatures: bool,
 }
 
 impl InputProcessor {
     /// Creates a new input processor.
-    pub fn new(include_nested_inputs: bool, show_expressions: bool, hide_defaults: bool) -> Self {
+    pub fn new(
+        include_nested_inputs: bool,
+        show_expressions: bool,
+        hide_defaults: bool,
+        type_signatures: bool,
+    ) -> Self {
         Self {
             results: Default::default(),
             include_nested_inputs,
             show_expressions,
             hide_defaults,
+            type_signatures,
         }
     }
 
@@ -296,13 +309,20 @@ impl InputProcessor {
                 Decl::Bound(decl) if !self.hide_defaults => {
                     let name = decl.name();
                     let expr = decl.expr();
+                    let key = namespace
+                        .clone()
+                        .push(name.text())
+                        .join()
+                        .expect("key to join");
 
-                    if let Some(value) = self.expression(&expr) {
+                    if self.type_signatures {
                         self.results
-                            .insert(namespace.clone().push(name.text()).join().unwrap(), value);
+                            .insert(key, Value::from(format!("{}", decl.ty())));
+                    } else if let Some(value) = self.expression(&expr) {
+                        self.results.insert(key, value);
                     } else if self.show_expressions {
                         self.results.insert(
-                            namespace.clone().push(name.text()).join().unwrap(),
+                            key,
                             Value::from(format!(
                                 "{ty} <NON-LITERAL: `{expr}`>",
                                 ty = decl.ty(),
@@ -314,26 +334,20 @@ impl InputProcessor {
                 Decl::Unbound(decl) => {
                     let name = decl.name();
                     let ty = decl.ty();
+                    let key = namespace
+                        .clone()
+                        .push(name.text())
+                        .join()
+                        .expect("key to join");
 
-                    if !ty.is_optional() {
+                    if self.type_signatures {
+                        self.results.insert(key, Value::from(format!("{}", ty)));
+                    } else if !ty.is_optional() {
                         // required input
-                        self.results.insert(
-                            namespace
-                                .clone()
-                                .push(name.text())
-                                .join()
-                                .expect("key to join"),
-                            Value::String(format!("{ty} <REQUIRED>")),
-                        );
+                        self.results
+                            .insert(key, Value::String(format!("{ty} <REQUIRED>")));
                     } else if !self.hide_defaults {
-                        self.results.insert(
-                            namespace
-                                .clone()
-                                .push(name.text())
-                                .join()
-                                .expect("key to join"),
-                            Value::Null,
-                        );
+                        self.results.insert(key, Value::Null);
                     }
                 }
                 _ => {
@@ -487,6 +501,7 @@ pub async fn inputs(args: Args, config: Config, colorize: bool) -> CommandResult
         args.nested_inputs,
         args.show_non_literals,
         args.hide_defaults,
+        args.type_signatures,
     );
 
     let ast = document