+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `--type-signatures` flag to `inputs` ([#1205](https://github.com/stjude-rust-labs/sprocket/pull/1205)).
+* `format check` notifies when the only diff is in newline style
+  ([#1204](https://github.com/stjude-rust-labs/sprocket/pull/1204)).
 * Added configuration setting `run.digest_cache_capacity` and
   `server.engine.digest_cache_capacity` for specifying the evaluation digest
   cache capacity ([#1178](https://github.com/stjude-rust-labs/sprocket/pull/1178)).
@@ -39,6 +42,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Input validation now identifies JSON and YAML input files that need an `@`
   prefix instead of reporting a misleading array type mismatch
   ([#906](https://github.com/stjude-rust-labs/sprocket/issues/906)).
+* Print an informative error when `dev test` `parallelism` argument or config
+  value is `0` instead of panicking ([#1196](https://github.com/stjude-rust-labs/sprocket/pull/1196)).
 
 ## 0.30.1 - 2026-08-27